First of a series of significant changes to how monitor is organized.    mailer.py -- uses CC rather than AdminCC to filter messages that are copied             to our support list.  Requires additional scrips in RT.     controllers_local.py -- supports local extensions to the web interface.

+++ monitor/wrapper/plc.py
@@ -412,7 +412,8 @@ def enableSiteSlices(loginbase):
 				for attr in l_attr:
 					if "enabled" == attr['tagname'] and attr['value'] == "0":
 						logger.info("Deleted enable=0 attribute from slice %s" % slice)
-						api.DeleteSliceTag(auth.auth, attr['slice_tag_id'])
+						if not isSliceExempt(slice):
+							api.DeleteSliceTag(auth.auth, attr['slice_tag_id'])
 		except Exception, exc:
 			logger.info("enableSiteSlices: %s" % exc)
 			print "exception: %s" % exc
